+++ src/api/resources/commons/errors/index.ts
@@ -1,5 +1,6 @@
 export * from "./ConflictError.js";
 export * from "./DoesNotExistError.js";
+export * from "./ForbiddenError.js";
 export * from "./InternalServerError.js";
 export * from "./InvalidRequest.js";
 export * from "./UnauthorizedError.js";